finance calculator

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 # this is basic finance calculator project based on one's total income and expenses in a month
-monthly_income = float(input("Enter your monthly income"))
-monthly_expenses = float(input("Enter your total monthly expenses"))
+monthly_income = int(input("Enter your monthly income"))
+monthly_expenses = int(input("Enter your total monthly expenses"))
 monthly_savings = monthly_income - monthly_expenses
 print(f"Your monthly savings  are, {monthly_savings}")
